refactor: decompose algorithm and runner into focused modules
Split the monolithic GeneticAlgorithm (870 lines) into BaseEngine ABC,
GeneticAlgorithm, StoppingCriteriaEvaluator, and FitnessCalculator.
Split KrknRunner into commands, composite, and orchestration modules.
This separation makes each file single-purpose and enables adding new
optimization engines (e.g. Optuna) by extending BaseEngine.
