+++ krkn_ai/algorithm/__init__.py
@@ -1,0 +1,4 @@
+from krkn_ai.algorithm.base import BaseEngine
+from krkn_ai.algorithm.genetic import GeneticAlgorithm
+
+__all__ = ["BaseEngine", "GeneticAlgorithm"]
