+++ krkn_ai/algorithm/genetic/__init__.py
@@ -1,0 +1,3 @@
+from krkn_ai.algorithm.genetic.engine import GeneticAlgorithm
+
+__all__ = ["GeneticAlgorithm"]
